requirements.txt
aliyun-python-sdk-core==2.13.36
aliyun-python-sdk-kms==2.16.1
amqp==5.1.1
asgiref==3.7.2
async-timeout==4.0.3
beautifulsoup4==4.12.2
billiard==4.1.0
celery==5.3.1
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
crcmod==1.7
cron-descriptor==1.4.0
cryptography==41.0.1
DingtalkChatbot==1.5.7
Django==4.2.2
django-bootstrap4==23.2
django-celery-beat==2.5.0
django-filter==23.2
django-grappelli==3.0.6
django-oss-storage==1.1.1
django-python3-ldap==0.15.4
django-redis==5.3.0
django-registration-redux==2.12
django-simpleui==2023.3.1
django-timezone-field==5.1
djangorestframework==3.14.0
dnspython==2.4.2
eventlet==0.33.3
flower==2.0.1
greenlet==2.0.2
gunicorn==21.2.0
h11==0.14.0
humanize==4.8.0
idna==3.4
jmespath==0.10.0
kombu==5.3.1
ldap3==2.9.1
Markdown==3.4.4
msgpack==1.0.5
mysqlclient==2.2.0
oss2==2.18.1
packaging==23.1
Pillow==10.0.0
prometheus-client==0.17.1
prompt-toolkit==3.0.39
pyasn1==0.4.8
pycparser==2.21
pycryptodome==3.18.0
python-crontab==3.0.0
python-dateutil==2.8.2
pytz==2023.3
redis==5.0.0
requests==2.31.0
sentry-sdk==1.29.2
six==1.16.0
soupsieve==2.4.1
sqlparse==0.4.4
tornado==6.3.3
typing_extensions==4.6.3
tzdata==2023.3
urllib3==2.0.3
uvicorn==0.23.2
vine==5.0.0
wcwidth==0.2.6